[ORA-698] [ORA-681] [ORA-678] Payment for Inference RequestInference Transaction Type (#19)
Creates a MsgServer handler to handle a new transaction type:
RequestInference. RequestInference takes an array of InferenceRequests
and processes them to make sure they're valid. If they are valid, it
transfers funds from the sender of the transaction in the amount
specified by the inference, and writes the inference into the
RequestInference mempool to be processed and eventually kick off
inferences by blockless.