Block proposer triggers inference/weight generation  (#46)
The goal of this PR is to make the block proposer the one that will
trigger the inference and weight generation for each topic. It will
solve the problem of each validator triggering it when we only need one.

To achieve that, I removed the inference/weight trigger from the`
EndBlock` inside the Module and added it in the `PrepareProposal` at the
App level, where the construction of the block is handled. You can set a
handler to be executed by the block proposer every block. Since the
block proposer has access to the mempool, it can also add additional
validation or inject new transactions, a feature we are not currently
using.

More about `PrepareProposal`:
https://docs.cosmos.network/main/build/abci/prepare-proposal