Implement functions for math formulas with unit tests (#83)
Implements equation 6 (phi), and 32-41 (Reputer intra-topic rewards)

Does not wire equations together. Just implements equations and provides
unit tests to make sure the math works as expected.

Un-fucks the integration tests in the major-upgrade repo so that tests
can actually pass.

Is built off an earlier commit in #79. At this point I would say it's
not dependent on 79, and the squash+merge will make 79 fine to merge
into major-upgrade (as long as guilherme pulls in changes)

---------

Co-authored-by: Guilherme Brandão <guilherme1237@gmail.com>