Fix ea 3-8 implementation and unit test (#120)
I found a couple of bugs in the implementation of
`CalcForcastImpliedInferences`. Fix those bugs. Note that only the first
test passes currently, the other two are pending a resolution of a
dispute over what the correct output value should be for the equation.
Diederik should be able to give final confirmation tomorrow morning.