[ORA-1166] Migrate float64 type to CockroachDB Arbitrary Precision Decimal type (#116)
float64 are handled differently on different CPU architecture so we
cannot use it for consensus when we expect different types of CPU
machines to come to consensus e.g. x86 and ARM

Fortunately underlying cockroachDB that cosmos-sdk already relies on
provides a complex arbitrary-precision decimal package for go called
`Dec.`

This PR migrates us to use `Dec` in place of float64.