Diego/ora 1174 add nonces and update b7s worker and reputer requests in (#119)
* Use Nonce as the blockHeight
* Use timestamp as param for reputers
* Activate TopicHandler
----
* Note 1: Several TODOs added - please check @kpeluso 

* Note 2: TopicHandler cannot call the fn on queryServer directly to get
inferences and pass them on to the fn.
The eq12 inference ValueBundle is calculated in the function
`queryServer.GetWorkerLatestInferenceByTopicId`.
This function is on queryServer, not on the keeper, which the
TopicHandler does not have access to.
This function should be on the Keeper. When attempting to internalize it
in the Keeper, there is a circular import dependency, because the
`inference_synthesis` module also import the keeper.

```
package github.com/allora-network/allora-chain/cmd/allorad
	imports github.com/allora-network/allora-chain/app
	imports github.com/allora-network/allora-chain/x/emissions/keeper
	imports github.com/allora-network/allora-chain/x/emissions/module/inference_synthesis
	imports github.com/allora-network/allora-chain/x/emissions/keeper: import cycle not allowed
```
Ig we should refactor those needs in inference_synthesis to remove the
import from the keeper, so we can safely call it from there. Passing the
params as dependencies instead of calling the functions inside.

* Note 3: In `queryServer.GetWorkerLatestInferenceByTopicId` [we return
5
fields](https://github.com/allora-network/allora-chain/blob/b6adcd0149f09e7fbcb8b3a6723889fb4f2cba69/x/emissions/module/inference_synthesis/network_inferences.go#L322),
not 7.
Shouldn't we also add the ones we derive from, ie inferences and
forecasts? We'll need them in the reputer anyway, so they can be put in
that format outside, but looks sensible to add them in there?

---------

Co-authored-by: Kenny <pelusoken@gmail.com>