Guilherme/ora 1212 create integration test for emissions module emit rewards (#174)
This is adding an integration test in the emissions module and the goal
is to simulate the happy path of our rewards cycle. It goes from topic
creation until rewards distribution.

To Consider:

- Rewards for workers are pending until the worker is fixed; right now,
they are receiving 0 rewards since the tax is higher than the reward.
- Added some TODOs. The main ones are related to deciding which block we
need to consider when querying for network losses; probably, it's the
latest one for all cases.
- The test is just ensuring that things are not breaking and rewards are
being generated, not checking values until we have the tax and the
reputers' scores calculation update (it will come in subsequent PRs).