Wrap errors to give more stack trace-like information (#190)
The EmitRewards function is not intended to panic on errors. This means
that the stack trace on errors in EmitRewards stops at the call of
EmitRewards and does not adequately help you determine where in the
rewards flow an error happened. I used my best judgement to avoid
wrapping absolutely every trivial error in e.g. an addition or
subtraction, but rather focusing on the function call stack.

This PR uses the cosmos errors.Wrapf function to help track down e.g.
consensus errors


it also tweaks some annoying formatting and adds a comment to a function
header for a function that lacked a function header comment