ORA1377 protect validators from egregious data (#231)
This is the straightforward solution that encompasses simply checking
the size of the input to query and msg server functions that could be
unbounded, bringing down a node. I found only two functions that can
have unbounded inputs because they take an array of some sort. I checked
the input size and throw an error if the input is too large, and chose a
size limit that allows for a few thousand items.

Perhaps controversial is I removed `extraData` from `MsgFundTopic` to
prevent arbitrarily large data from being sent there. It is unused, but
has a comment saying that it *might* be used in the future.

---------

Signed-off-by: br4e <milesminton@gmail.com>
Co-authored-by: Kenny P <17100641+kpeluso@users.noreply.github.com>