[ORA-1271] Convert units of Stake to cosmosMath.Int (from Uint) (because sdk.coins are in Int) (#279)
Back in December of last year when we were first building the allora
staking code for the reputers and workers, I implemented stake with
cosmosMath.Uint because it was my estimation that stake values should
never be negative (this is still true). But in practice, stake values
are in units of sdk.Coins, which are tokens, and the underlying
cosmos-sdk uses cosmosMath.Int for coins. In order to make the code
cleaner and cut down on a lot of
`cosmosMath.NewIntFromBigInt(existingUint.ToBigInt())` calls, we simply
replace the staking units to be in cosmosMath.Int directly.

---------

Co-authored-by: Kenny <pelusoken@gmail.com>