Fix TestTotalInferersRewardFractionGrowsWithMoreInferers (#304)
`TestTotalInferersRewardFractionGrowsWithMoreInferers` was not testing
to see if the correct ratios (inferer and forecaster reward fractions)
were increasing with increasing inferer and forecasters, resp. It was
testing a completely different ratio not of concern.

Additionally, this test was flaky in that it would usually pass but only
sometimes fail. This was due to a randomness seeded with a sensitive
value (nanosecond time).

That volatility meant that extra inferers could have high losses => they
actually may not earn a greater fraction, because the protocol will pay
classes of actors out more based on their numbers but also their
performance. so the test failed because ~20% of the times their losses
were so high that despite their greater number => they earned less,
which is actually expected. I solved this in an expedient way, the
better way would be for me to care about how high the losses were before
i decide what to test, the best way would be for me to systematically
adjust that loss, but not of immediate concern