ORA1554 Thoroughly test revamp synth pr (#334)
This is _so close_. All of the test values are doing the thing where
they are basically right, except they are about 5% off of their correct
values.

I have noticed that the Forecast implied inferences are a little bit
off, and that is throwing off all of the other values, EXCEPT for the
naive network inference which is exactly correct, _which does not use
forecast implied inferences_

If we can just figure out why forecast implied inferences are not
returning the correct value, I think we've got it.

I've looked at it a few times. Anyone else want to give it a look? You
can run `TestForecastImpliedInferences` to test.