Add nil dereference coverage and error control (#348)
Add coverage against unexpected nil pointer dereferences.
It will throw error in that case.

Note: This is a very conservative take here, sending error if there is
any - thus sending no weights; another option could be to ignore those
that are nil but still return the completed ones, if any. But I'm unsure
of the consequences, so rather return error and handle as such.