Check Invariants Hold During Integration Testing (#393)
Check the following invariants during the course of an integration test
run:

* The balance of the totalStake in the emissions keeper should always
equal the cosmos-sdk /x/bank balance of the "allorastaking" bank account
* The sum of all topicStake in every topic should always equal the
totalStake
* The sum of all stakeReputerAuthority for a given topic should always
equal the topic stake for that topic
* The sum of all stakeReputerAuthority should equal the total stake
* The stakeSumFromDelegator should always equal the sum of
`delegatedStakes[topicid, delegator, all reputers for that delegator])`
* The stakeFromDelegatorsUponReputer should always equal the sum of
`delegatedStakes[topicid, all delegators for that reputer, reputer]`
* The sum of all (delegateRewardsPerShare * delegated stake - reward
debt) = the balance of the /x/bank
AlloraPendingRewardForDelegatorAccountName module account
* The state of the stakeRemovalsByBlock and stakeRemovalsByActor should
always be identical in terms of the number of keys in the map, and the
content of those keys
* The sum of all stakeRemovals should always be less than or equal to
the totalStake


The invariant checking lives in a file called invariants.go. The cosmos
/x/invariants module apparently is not functional at this time so rather
than rely on their module, we just edit the ABCI endBlock to check the
invariants at the end of every block. This is not perfect, but better
than nothing. The code to do that is stored in a patch that is applied
in the github workflow via `git apply ./test/invariant/invariants.patch`