EpochLength changes + Using last commit at request time (#412)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v                               ✰  Thanks for creating a PR! ✰
v    Before smashing the submit button please review the checkboxes.
v If a checkbox is n/a - please still include it but + a little note why
v If your PR doesn't close an issue, that's OK! Just remove the Closes:
#XXX line!
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

Closes: #XXX

## What is the purpose of the change

* Defines worker nonce as blockHeight, instead of blockheight +
epochLength. Meaning is changed from "the block for when the inference
should be made" to "the block when the request is made".
* In places where this is used, adjust adding epochLength where
appropriate

* Use last reputer commit nonce as previousLosses value where
applicable. Keep previous value (`reputerNonce - epochLength`) on first
commit and as default if no losses are found.

## Testing and Verifying

*(Please pick one of the following options)*

This change added tests and can be verified as follows:

*(example:)*
  - *Modifies the topNonces tests that are failing due to the changes*

## Documentation and Release Note

- [ ] Does this pull request introduce a new feature or user-facing
behavior changes?


Where is the change documented?
  - [ ] Specification (`x/{module}/README.md`)
- [ ] Allora documentation site `docs.allora.network` source code at:
`https://github.com/allora-network/docs`
  - [ ] Code comments?
  - [ X ] N/A

Will be documented as part of a larger change on nonces management.