Batch Stake Removals in EndBlock so attackers cannot spam chain to a slow halt (#469)
## What is the purpose of the change

The EndBlock can slow down the consensus process for the blockchain
causing block production times to go up, if EndBlock does too much
computation.

This PR puts a limit on the number of stake removals that can be
processed every block.

We just do a reasonable limit every block until we clear out the backlog
- this new parameter `half_max_process_stake_removals_end_block` is
tunable in the moduleParams

## Testing and Verifying

Our existing tests should at least confirm that stake removals function
the same as they did before under non-hostile conditions

I have added the following to test the new "limit" feature on getting
block removals.
* `TestGetStakeRemovalsUpUntilBlockLimitLessThanNumRemovals`
* `TestGetStakeRemovalsUpUntilBlockLimitGreaterThanNumRemovals`
* `TestGetStakeRemovalsUpUntilBlockLimitReturnsPreviousBlocks`

## Documentation and Release Note

@kush-alloralabs we should probably make it clear in the documentation
that stake removals are no longer guaranteed to be processed exactly
`RemovalDelayWindow` long blocks. It could be longer if there's a big
queue of removals to process. In that case we will do
`HalfMaxProcessStakeRemovalsEndBlock` number of removals each block
until the queue is done