[ORA-1613] Initial Value for emission per unit staked token = target (#486)
## What is the purpose of the change

We calculate target emission and the emission rate on a monthly cadence.
In the first month we have no previous history to use for the
exponential moving average which returns e_i. Therefore, in that case,
we use the target e_i instead of the exponential moving average,
directly.

## Testing and Verifying

Our existing tests should at least cover that this does not break
anything

## Documentation and Release Note

No Documentation changes needed.