Fix staticcheck Linter Errors (#500)
This change goes through and fixes bugs discovered by systematic running
of the golang `staticcheck` tool.

## Testing and Verifying

This PR will be tested by our existing test suite, and / or is an
addendum to our existing test suite. It is not meant to change the
functionality of the chain or introduce new features.

**REVIEWER NOTE:** I commented out the unused functions
`scheduleUpgrades` in `app/upgrades.go` and `removeOldKVStores` in
`x/emissions/migrations/v2/migrate.go` - I don't think either are
necessary or correct for v0.3.0 release but I kept them commented out in
case they're needed for reference. If that was the wrong thing to do,
please lets figure out a better solution.

## Documentation and Release Note

No documentation implications.

---------

Co-authored-by: Guilherme Brandão <guilherme1237@gmail.com>