Upgrade process improvement (#520)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v           ✰  Thanks for creating a PR! You're awesome! ✰
v Please note that maintainers will only review those PRs with a
completed PR template.
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Purpose of Changes and their Description
To solidify the process of upgrading with docker. 
Now provide an image with tag appended with -upgrader

## Are these changes tested and documented?

Yes tested manually. Downlaoded snapshot and then tested with the
snapshot and it upgraded as expected

---------

Signed-off-by: Tobi Okedeji <tobiokedeji@gmail.com>