Undo overwriting of protobufs since v0.3.0 (#530)
## Purpose of Changes and their Description

#528 and others overwite the protobuf in the v2 folder of the emissions
module. this is the incorrect way to perform software upgrades.

Future changes to protobufs must go in their own v3 folder and the v2
must stay pinned to v0.3.0

This PR reverts the changes and introduces a new v3 folder for working
with.

## Link(s) to Ticket(s) or Issue(s) resolved by this PR

No written linear tasks.

## Are these changes tested and documented?

Covered by existing unit tests.

## Still Left Todo

Migrations for the new parameters in #528. I already wrote migration
code in #525 and #529 so I recommend either basing off of those or
waiting for at a miniumum, #525 to be merged.