Delete mint migration for version 2, since we didn't use it on testnet (#541)
## Purpose of Changes and their Description

Due to errors with the mint module `ConsensusVersion`, when we upgraded
to v0.3.0, we did not actually run the mint module migration that we
intended to run. In PROTO-2136, we addressed this by doing a manual
parameter upgrade on chain via transaction. So it ended up not being an
issue that the migration never ran.

In order to reduce confusion, and more importantly _to prevent the
migration from accidentally running in the future_ when we do an upgrade
to v0.4.0, this PR deletes the migration code that was never used.

## Link(s) to Ticket(s) or Issue(s) resolved by this PR

not sure, maybe @fernandofcampos has a ticket related to this

## Are these changes tested and documented?

This PR deletes code and so needs no documentation or tests