simplify util_sort.go FindTopNByScoreDesc (#539)
## Purpose of Changes and their Description

As part of #421 I need to add features to util_sort for getting the top
list of actors by score. I also observed some ways to make the
`FindTopNByScoreDesc` function simpler as part of that effort. But due
to the high level of importance of this code I wanted to do a refactor
of this function in an isolated manner to make sure that no changes that
affect tests / state are made by the way I rewrote the function. Hence
this PR

## Link(s) to Ticket(s) or Issue(s) resolved by this PR

Related to PROTO-2204

## Are these changes tested and documented?

This code is a refactor and so it is tested by our existing unit tests
to ensure that no logical changes have been made.