Bump Changelog Version (#546)
## Purpose of Changes and their Description

Our changelog is out of date on dev branch. This sets all the changes
that were v0.3.0 as marked for v0.3.0 and starts a new category for
v0.4.0.

## Future Work / Todo

We should avoid accidentally updating the wrong version's changelog with
more formal processes. We should also require changes to changelog for
every PR, maybe enforced by linter rules on the changelog.