Adjust mint module to use v1beta1 protobuf imports in v2 protobuf (#589)
## Purpose of Changes and their Description

I figured out how to share protobuf from a previous module version, so
you don't have to copy every single protobuf over when you edit a
protobuf file for an upgrade. This does that, for the mint module. Now
the only new file in the x/mint/proto/v4 folder is just the tx.proto
file - since that's the only proto that needs to change for 0.4.1

## Are these changes tested and documented?

This is a refactor that does not change any code functionality