Clear NaN in maps by deleting all map values during migration (#590)
## Purpose of Changes and their Description

This PR enforces the "clearing of poison" by removing all not a number
(NaN) values in the chain. It does so by completely and totally deleting
the contents of the inference scores, forecaster scores, emas, loss
bundles etc maps.

## Are these changes tested and documented?

Tested in `TestResetMapsWithNonNumericValues`

No documentation, no changes to change log as this is part of #579.