Rename v0.4.1 to v0.5.0 (#594)
## Purpose of Changes and their Description

We changed our minds about the next upgrade being a minor version and
instead we're bumping the name to a major version.

Therefore this PR makes the necessary changes inside the codebase to
ensure that we're naming this next upgrade correctly

## Are these changes tested and documented?

This PR is only a refactoring change, no actual functionality is
changed.