Math Operations on NaNs Should Return Errors (#598)
## Purpose of Changes and their Description

This PR introduces checks in the Allora math library when operating on
NaN values. This PR makes it so that on addition, subtraction, `.Abs()`,
etc of a value with a NaN, the addition or subtraction etc function
returns an error.

This PR also takes a coverage based approach to write tests for many
math functions that previously did not have coverage.

This PR also refactors various functions in the math package.

## Link(s) to Ticket(s) or Issue(s) resolved by this PR

PROTO-2315

## Are these changes tested and documented?

These changes are tested by our existing unit tests. In addition, I have
added tests to cover the new error cases when a math function ought to
error.

Documentation changes are not required since alloraMath is an
internally consumed library.