Validate all Values for Correctness Prior to Storing them in the Keeper
In order to aid in ensuring correct state, and to practice defense-in-depth, we should check all values for correctness at the final step, right before the values are written into the keeper.

This code moves all the separate validation functions in the `types` folder into one `validations.go` file which is easier for someone looking for validations to go look for, rather than trying to remember the name of the type being validated.

It modifies all of the various keeper.go functions that do keeper.xyz.Set() functions. I went through and manually checked every single field in the keeper for where it's set and set up a validator prior to that line.

PROTO-2341, all of the NaN work in release v0.5.0

This code is tested by our existing unit tests, which proceed to fall over and took a week of rewriting to make pass.

This PR does not change any user or external facing code, so no documentation updates are needed.

PR added to the `Unreleased` section of `CHANGELOG.md`

Unit tests are in #608 just to keep the PR diff smaller for reviewers.

PROTO-2369 was created by me to track one place in the code where we do incomplete validation in this PR: we modify loss bundles at the filtering step in CloseReputerNonce. This causes signature verification on the bundles to fail when it comes time to insert them into the database. The correct solution for this is to change the data type that we are storing these modified filtered loss bundles in. For now that can stay as future work.