Allow forecasts with unregistered inferers to proceed (#777)
<!-- < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < < ☺
v           ✰  Thanks for creating a PR! You're awesome! ✰
v Please note that maintainers will only review those PRs with a
completed PR template.
☺ > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > > >  -->

## Purpose of Changes and their Description

This addresses the vulnerability where inferer deregistrations would
cause a forecaster's entire submission to fail. Now, when an inferer is
not registered in a topic, we skip that specific inferer in the forecast
rather than failing the whole transaction.

- Modified message handler to continue processing when an inferer is
unregistered
- Added test case to verify the fix works correctly

## Link(s) to Ticket(s) or Issue(s) resolved by this PR


https://linear.app/alloralabs/issue/ENGN-3519/172-inferer-de-registrations-will-cause-forecasting-submission-to

## Are these changes tested and documented?

- [x] If tested, please describe how. If not, why tests are not needed.
- [x] If documented, please describe where. If not, describe why docs
are not needed.
- [x] Added to `Unreleased` section of `CHANGELOG.md`?

## Still Left Todo

*Fill this out if this is a Draft PR so others can help.*